Fixes and improvements to parsing of `static replaces`
There was previously a mismatch between our parsing and runtime where it was assumed that `typealias ReplacedAssembly` would be appended to a manually declared `static replaces`, but that is not possible with default extensions. So now the parsing has been updated to correctly reflect the runtime. Additionally there is now an IDE error produced by Knit if a manual `static replaces` is declared but it misses the `typealias ReplacedAssembly` type.