Fix treatment of `@escaping` attribute on registration arguments and fix identifier uniqueness bug
Previously we were dropping `@escaping` when parsing the argument type declaration, but then adding it back upon writing the type safety source file. However this meant that an attributed typealias such as `@escaping ClosureActionTypeAlias` would drop the attribute on parsing, but we would have no way to know to add it back upon writing the type safety methods.

Now we do not make any modifications to the attributes on type declarations for arguments. When generating properties for those types, `@escaping` is always dropped as it is implicit (and including it explicitly is a compilation error).

This also uncovered an issue with argument identifier uniqueness for closures. We should use the identifier when determining uniqueness, not the type, as all closure types map to the same identifier.