Allow for auto configuration of containers when using ModuleAssembler
When using a single ModuleAssembler it can be useful for it to ignore TargetResolvers for its assemblies.

The app target use case is an app that only has a single TargetResolver, often `Knit.Resolver`.

The test target use case is when a single ModuleAssembler is used for multiple TargetResolvers for simplicity of test setup.